==> R/04_PlotsRevised.R <==
##revised plots
library(rnaturalearth)
library(rnaturalearthdata)
library(gridExtra)
library(sf)
library(patchwork)
library(gmRi)

##use rev_data from 03_Revisions.R
rev_data<-readRDS(here("Data", "revised_plot_data.RDS"))

#average lat####
nrow(rev_data)
lat<-vector("list", length = 49)
names(lat)=paste(unique(rev_data$comname))

for(i in 1:49){
  print(i)
  loop_df<-rev_data[i,]%>%
    unnest(data)%>%
    select(comname, est_year, avg_lat)%>%
    group_by(comname)
  
  group1<-rev_data[i,]%>%
    unnest(data)%>%
    filter(est_year %in% c(1970:2009))%>%
    group_by(comname)%>%
    nest()%>%
    mutate(overall_lat = map(data, possibly(mean_lat, NA)))%>%
    unnest(data)
  
  group2<-rev_data[i,]%>%
    unnest(data)%>%
    filter(est_year %in% c(2010:2019))%>%
    group_by(comname)%>%
    nest()%>%
    mutate(overall_lat = map(data, possibly(mean_lat, NA)))%>%
    unnest(data)
  
  lat[[i]]<- ggplot(data=loop_df, aes(x=est_year, y=avg_lat))+
    geom_line(color = "#E9E9E9", linewidth = 0.5)+
    geom_point(size=0.5)+
    theme_gmri(axis.title = element_blank(),
               plot.title = element_text(size = 11),
               axis.text.y = element_text(size=10))+
    ggtitle(toupper(names(lat)[i]))+
    geom_line(data=group1, aes(x=est_year, y=as.numeric(overall_lat)), color="#00608A", linewidth=0.75)+
    geom_line(data=group2, aes(x=est_year, y=as.numeric(overall_lat)), color="#EA4F12", linewidth=0.75)
}
print(lat[2])

#average lon####
lon<-vector("list", length = 49)
names(lon)=paste(unique(rev_data$comname))

for(i in 1:49){
  print(i)
  loop_df<-rev_data[i,]%>%
    unnest(data)%>%
    select(comname, est_year, avg_lon)%>%
    group_by(comname)
  
  group1<-rev_data[i,]%>%
    unnest(data)%>%
    filter(est_year %in% c(1970:2009))%>%
    group_by(comname)%>%
    nest()%>%
    mutate(overall_lon = map(data, possibly(mean_lon, NA)))%>%
    unnest(data)
  
  group2<-rev_data[i,]%>%
    unnest(data)%>%
    filter(est_year %in% c(2010:2019))%>%
    group_by(comname)%>%
    nest()%>%
    mutate(overall_lon = map(data, possibly(mean_lon, NA)))%>%
    unnest(data)
  
  lon[[i]]<- ggplot(data=loop_df, aes(x=est_year, y=avg_lon))+
    geom_line(color = "#E9E9E9", linewidth = 0.5)+
    geom_point(size=0.5)+
    theme_gmri(axis.title = element_blank(),
               plot.title = element_text(size = 11),
               axis.text.y = element_text(size=10))+
    ggtitle(toupper(names(lon)[i]))+
    geom_line(data=group1, aes(x=est_year, y=as.numeric(overall_lon)), color="#00608A", linewidth=0.75)+
    geom_line(data=group2, aes(x=est_year, y=as.numeric(overall_lon)), color="#EA4F12", linewidth=0.75)
}
print(lon[2])

##average surface temp
sst<-vector("list", length=49)
names(sst)=paste(unique(rev_data$comname))

for(i in 1:49){
  print(i)
  loop_df<-rev_data[i,]%>%
    unnest(data)%>%
    select(comname, est_year, avg_sur_temp)%>%
    group_by(comname)
  
  group1<-rev_data[i,]%>%
    unnest(data)%>%
    filter(est_year %in% c(1970:2009))%>%
    group_by(comname)%>%
    nest()%>%
    mutate(overall_sst = map(data, possibly(avg_sst, NA)))%>%
    unnest(data)
  
  group2<-rev_data[i,]%>%
    unnest(data)%>%
    filter(est_year %in% c(2010:2019))%>%
    group_by(comname)%>%
    nest()%>%
    mutate(overall_sst = map(data, possibly(avg_sst, NA)))%>%
    unnest(data)
  
  sst[[i]]<- ggplot(data=loop_df, aes(x=est_year, y=avg_sur_temp))+
    geom_line(color = "#E9E9E9", linewidth = 0.5)+
    geom_point(size=0.5)+
    theme_gmri(axis.title = element_blank(),
               plot.title = element_text(size = 11),
               axis.text.y = element_text(size=10))+
    ggtitle(toupper(names(sst)[i]))+
    geom_line(data=group1, aes(x=est_year, y=as.numeric(overall_sst)), color="#00608A", linewidth=0.75)+
    geom_line(data=group2, aes(x=est_year, y=as.numeric(overall_sst)), color="#EA4F12", linewidth=0.75)
}

#bottom temp####
bt<-vector("list", length=49)
names(bt)=paste(unique(rev_data$comname))

for(i in 1:49){
  print(i)
  loop_df<-rev_data[i,]%>%
    unnest(data)%>%
    select(comname, est_year, avg_bot_temp)%>%
    group_by(comname)
  
  group1<-rev_data[i,]%>%
    unnest(data)%>%
    filter(est_year %in% c(1970:2009))%>%
    group_by(comname)%>%
    nest()%>%
    mutate(overall_bt = map(data, possibly(avg_bt, NA)))%>%
    unnest(data)
  
  group2<-rev_data[i,]%>%
    unnest(data)%>%
    filter(est_year %in% c(2010:2019))%>%
    group_by(comname)%>%
    nest()%>%
    mutate(overall_bt = map(data, possibly(avg_bt, NA)))%>%
    unnest(data)
  
  bt[[i]]<- ggplot(data=loop_df, aes(x=est_year, y=avg_bot_temp))+
    geom_line(color = "#E9E9E9", linewidth = 0.5)+
    geom_point(size=0.5)+
    theme_gmri(axis.title = element_blank(),
               plot.title = element_text(size = 11),
               axis.text.y = element_text(size=10))+
    ggtitle(toupper(names(bt)[i]))+
    geom_line(data=group1, aes(x=est_year, y=as.numeric(overall_bt)), color="#00608A", linewidth=0.75)+
    geom_line(data=group2, aes(x=est_year, y=as.numeric(overall_bt)), color="#EA4F12", linewidth=0.75)
}


#depth###
depth<-vector("list", length = 49)
names(depth)=paste(unique(rev_data$comname))

for(i in 1:49){
  print(i)
  loop_df<-rev_data[i,]%>%
    unnest(data)%>%
    select(comname, est_year, avg_depth)%>%
    group_by(comname)
  
  group1<-rev_data[i,]%>%
    unnest(data)%>%
    filter(est_year %in% c(1970:2009))%>%
    group_by(comname)%>%
    nest()%>%
    mutate(overall_depth = map(data, possibly(mean_depth, NA)))%>%
    unnest(data)
  
  group2<-rev_data[i,]%>%
    unnest(data)%>%
    filter(est_year %in% c(2010:2019))%>%
    group_by(comname)%>%
    nest()%>%
    mutate(overall_depth = map(data, possibly(mean_depth, NA)))%>%
    unnest(data)
  
  depth[[i]]<- ggplot(data=loop_df, aes(x=est_year, y=avg_depth))+
    geom_line(color = "#E9E9E9", linewidth = 0.5)+
    geom_point(size=0.5)+
    theme_gmri(axis.title = element_blank(),
               plot.title = element_text(size = 11),
               axis.text.y = element_text(size=10))+
    ggtitle(toupper(names(depth)[i]))+
    scale_y_reverse()+
    geom_line(data=group1, aes(x=est_year, y=as.numeric(overall_depth)), color="#00608A", linewidth=0.75)+
    geom_line(data=group2, aes(x=est_year, y=as.numeric(overall_depth)), color="#EA4F12", linewidth=0.75)
}

###print plots alphabetically in multipage layouts
lat_list   <-   lat[c(1:49)]
lon_list   <-   lon[c(1:49)]
sst_list   <-   sst[c(1:49)]
bt_list    <-    bt[c(1:49)]
depth_list <- depth[c(1:49)]

##lat plots
species_lat<-marrangeGrob(lat_list, layout_matrix = matrix(1:20,  nrow = 5, ncol=4, byrow=TRUE), top=NULL)
ggsave("lat_multipanel_v3.pdf", species_lat, height = 15, width = 12.5, units ="in")

#lon plots
species_lon<-marrangeGrob(lon_list, layout_matrix = matrix(1:20,  nrow = 5, ncol=4, byrow=TRUE), top=NULL)
ggsave("lon_multipanel_v3.pdf", species_lon, height = 15, width = 12.5, units ="in")

#sst plots
species_sst<-marrangeGrob(sst_list, layout_matrix = matrix(1:20,  nrow = 5, ncol=4, byrow=TRUE), top=NULL)
ggsave("sst_multipanel_v3.pdf", species_sst, height = 15, width = 12.5, units ="in")

#bt plots
species_bt<-marrangeGrob(bt_list, layout_matrix = matrix(1:20,  nrow = 5, ncol=4, byrow=TRUE), top=NULL)
ggsave("bt_multipanel_v3.pdf", species_bt, height = 15, width = 12.5, units ="in")

#depth plots
species_depth<-marrangeGrob(depth_list, layout_matrix = matrix(1:20,  nrow = 5, ncol=4, byrow=TRUE), top=NULL)
ggsave("depth_multipanel_v3.pdf", species_depth, height = 15, width = 12.5, units ="in")

#MOVERS MAP
movers<-read_csv("Data/movers.csv", col_names=FALSE)
movers<-movers%>%
  rename("comname" = "X1")
cap_names<-capitalize(movers$comname)
legend<-movers%>%
  cbind(cap_names)%>%
  select(cap_names)%>%
  rename("Common Name" = "cap_names")

#map legend 
legend<-ggtexttable((legend%>%
                       distinct()),
                    theme= ttheme(base_size = 8))
print(legend)
ggsave("legend.png", legend)

#directions
n_movers<-group_2_rev%>%
  select(comname, mean_lat, mean_lon, group)%>%
  rbind(group_1_rev %>%
          select(comname, mean_lat, mean_lon, group))%>%
  filter(comname %in% movers$comname)%>%
  filter(!comname %in% c("cusk", "little skate", "windowpane"))%>%
  mutate(Direction = "North")

s_movers<-group_2_rev%>%
  select(comname, mean_lat, mean_lon, group)%>%
  rbind(group_1_rev %>%
          select(comname, mean_lat, mean_lon, group))%>%
  filter(comname %in% c("cusk", "little skate",
                        "windowpane"))%>%
  mutate(Direction = "South")

pairs<-rep(1:(44/2), each=2)

movers<-n_movers%>%
  rbind(s_movers)%>%
  arrange(comname)%>%
  cbind(pairs)

movers<-movers%>%
  rename("pairs" = "...6")

mover_names<-movers%>%
  filter(group == "group_2")

cap_names<-capitalize(mover_names$comname)

legend<-mover_names%>%
  cross_join(cap_names, copy=TRUE)%>%
  select(y, pairs)%>%
  rename("Common Name" = "y")%>%
  select(!comname)

#map
world <- ne_countries(scale = "medium", returnclass = "sf")
movement<-ggplot(data=world)+
  geom_sf()+
  coord_sf(xlim=c(-76, -66), ylim=c(37.5,47))+
  geom_line(data=movers, aes(x=mean_lon, y=mean_lat, group = comname), color="#535353", linewidth = 0.5)+
  geom_point(data=movers, aes(x=mean_lon, y=mean_lat, color= group), size=2)+
  theme_gmri()+
  scale_color_manual(name = "Years", labels = c("1970-2009", "2010-2019"), values=c("#00608A","#EA4f12"))+
  ggtitle("Mean Center of Biomass")+
  ylab("Center of Latitude")+
  xlab("Center of Longitude")+
  geom_text(data=mover_names, aes(x=mean_lon, y=mean_lat, label=pairs))+
  scale_y_continuous(breaks = c(36,40,44)) + scale_x_continuous(breaks = c(-78,-72,-66))
print(movement)

movement_no_labels<-ggplot(data=world)+
  geom_sf()+
  coord_sf(xlim=c(-76, -66), ylim=c(37.5,47))+
  geom_line(data=movers, aes(x=mean_lon, y=mean_lat, group = comname), color="#535353", linewidth = 0.5)+
  geom_point(data=movers, aes(x=mean_lon, y=mean_lat, color= group), size=2)+
  theme_gmri()+
  scale_color_manual(name = "Years", labels = c("1970-2009", "2010-2019"), values=c("#00608A","#EA4f12"))+
  ggtitle("Mean Center of Biomass")+
  ylab("Center of Latitude")+
  xlab("Center of Longitude")+
  scale_y_continuous(breaks = c(36,40,44)) + scale_x_continuous(breaks = c(-78,-72,-66))
print(movement_no_labels)

ggsave("movement_noLabels.png", movement_no_labels, height = 10, width = 10, units="in")

#grid.arrange(movement, legend, ncol = 3, layout_matrix =cbind(1,1,2))

#species decadal maps
map_data<-rev_data%>%
  unnest(data)%>%
  select(comname, est_year, data)%>%
  unnest(data)%>%
  group_by(comname)

#maps####
#with season
map_df<-map_data%>%
  group_by(comname)%>%
  nest()
nrow(map_df)
maps<-vector("list",length=49)
names(maps)=paste(unique(map_df$comname))


for(i in 1:49){
  print(i)
  loop_df<-map_df[i,]%>%
    unnest(data)%>%
    group_by(comname)
  
  maps[[i]]<-ggplot(data=world)+
    geom_sf()+
    coord_sf(xlim=c(-80, -65), ylim=c(30,47))+
    geom_point(data=loop_df, aes(x=avg_lon,y=avg_lat, color=season))+
    theme_gmri()+
    ggtitle(toupper(names(maps)[i]))+
    ylab("Center of Latitude")+
    xlab("Center of Longitude")+
    scale_y_continuous(breaks = c(36,40,44)) + scale_x_continuous(breaks = c(-78,-72,-66)) +
    facet_wrap(~decade, ncol=5)
}

species_maps<-marrangeGrob(maps, layout_matrix = matrix(1:3,  nrow = 3, ncol=1, byrow=TRUE), top=NULL)
ggsave("species_maps.pdf", species_maps, height = 15, width = 12.5, units ="in")

#season averaged over year
decade_maps<-vector("list",length=49)
names(decade_maps)=paste(unique(rev_data$comname))

for(i in 1:49){
  print(i)
  loop_df<-rev_data[i,]%>%
    unnest(data)%>%
    group_by(comname)%>%
    mutate(decade =10*est_year %/% 10)
  
  decade_maps[[i]]<-ggplot(data=world)+
    geom_sf()+
    coord_sf(xlim=c(-80, -65), ylim=c(30,47))+
    geom_point(data=loop_df, aes(x=avg_lon,y=avg_lat), color="#00736D", size = 0.75)+
    theme_gmri()+
    ggtitle(toupper(names(maps)[i]))+
    ylab("Center of Latitude")+
    xlab("Center of Longitude")+
    scale_y_continuous(breaks = c(36,40,44)) + scale_x_continuous(breaks = c(-78,-72,-66)) +
    facet_wrap(~decade, ncol=5)
}
decade_species_maps<-marrangeGrob(decade_maps, layout_matrix = matrix(1:4,  nrow = 4, ncol=1, byrow=TRUE), top=NULL)
ggsave("decade_species_maps.pdf", decade_species_maps, height = 15, width = 12.5, units ="in")
